@@ -1,11 +1,10 @@
 package me.xjqsh.lesraisinsarmor.mixin;
 
 import me.xjqsh.lesraisinsarmor.effect.SuitEffect;
-import net.minecraft.potion.Effect;
-import net.minecraft.potion.EffectInstance;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -13,15 +12,14 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @OnlyIn(Dist.CLIENT)
-@Mixin(EffectInstance.class)
+@Mixin(MobEffectInstance.class)
 public abstract class EffectInstanceMixin {
-    @Shadow public abstract Effect getEffect();
 
-    @Shadow @Final private Effect effect;
+    @Shadow public abstract MobEffect getEffect();
 
-    @Inject(method = "isNoCounter", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "isInfiniteDuration", at = @At("HEAD"), cancellable = true)
     public void isNoCounter(CallbackInfoReturnable<Boolean> cir) {
-        if(this.effect instanceof SuitEffect) {
+        if(this.getEffect() instanceof SuitEffect) {
             cir.setReturnValue(true);
             cir.cancel();
         }
